+++ scripts/run-cooler.sh
@@ -3,8 +3,9 @@ pairs_file=$1  # bgzipped, with .px2 index
 chrsize_file=$2
 bin_size=$3
 max_iter=$4 # e.g. 3000 (default 200)
-ncores=$5 # (default 8)
-out_prefix=$6
+ncores=$4 # (default 8)
+out_prefix=$5
+
 
 cp $chrsize_file ./tmpchrsize
 
@@ -15,5 +16,7 @@ cp $chrsize_file ./tmpchrsize
 cooler cload pairix -p $ncores ./tmpchrsize:$bin_size $pairs_file $out_prefix.cool
 
 # matrix balancing
-cooler balance --max-iters $max_iter $out_prefix.cool
-
+if [[ $balance == '1' ]]
+do
+  cooler balance --max-iters $max_iter $out_prefix.cool
+fi